ProgressTracker: fix bugs on .gz files.
- .gz files cannot be both read- and writable at the same time. In order
to support .gz files (and other file formats that do not support reading
and writing to the same file handle, the ReadAppendFile tuple is introduced.
This tuple contains methods 'open_read', and 'open_write', which take no
arguments and should open an underlying file to read from and write to,
respectively.